move experimental APIs to Schema.ts
